Move options-related things to another file
