+++ src/index.ts
@@ -1,8 +1,7 @@
 import { stringifiedRequestAttributes } from './formatting'
-import { VariableFilter } from './formatting'
-
+import { defaultOptions, Options, UserOptions } from './options'
 export default class ApolloLogExtension {
-  
+
   private options: Options
 
   constructor(ops: UserOptions = {}) {
@@ -36,38 +35,4 @@ export default class ApolloLogExtension {
   private log(msg: string): void {
     this.options.logger.log(`${this.options.prefix()} ${msg}`)
   }
-}
-
-const defaultOptions = {
-  logger: console,
-  logRequests: true,
-  logResponses: false,
-  prefix: () => `[${Date.now()}]`,
-  variableFilter: {
-    keywords: ['password'],
-    replacementText: '[FILTERED]'
-  },
-  ignoreSchemaRequest: false
-}
-
-interface Options {
-  logger: Logger
-  logRequests: boolean
-  logResponses: boolean
-  prefix: () => string
-  variableFilter: VariableFilter | false
-  ignoreSchemaRequest: boolean
-}
-
-export interface UserOptions {
-  logger?: Logger
-  logRequests?: boolean
-  logResponses?: boolean
-  prefix?: () => string
-  variableFilter?: VariableFilter | false
-  ignoreSchemaRequest?: boolean
-}
-
-export interface Logger {
-  log(msg: string): any
 }
